♻️ extract eventsUtils
