+++ packages/rum/src/domain/record/observers/inputObserver.ts
@@ -2,9 +2,9 @@ import type { DefaultPrivacyLevel, ListenerHandler } from '@datadog/browser-core
 import { instrumentSetter, assign, DOM_EVENT, addEventListeners, forEach } from '@datadog/browser-core'
 import { NodePrivacyLevel } from '../../../constants'
 import type { InputState } from '../../../types'
+import { getEventTarget } from '../eventsUtils'
 import { getNodePrivacyLevel, shouldMaskNode } from '../privacy'
 import { getElementInputValue, getSerializedNodeId, hasSerializedNode } from '../serialization'
-import { getEventTarget } from './utils'
 
 type InputObserverOptions = {
   domEvents?: Array<DOM_EVENT.INPUT | DOM_EVENT.CHANGE>
