--- packages/rum/src/domain/record/utils.ts
@@ -1,3 +1,0 @@
-export function isTouchEvent(event: MouseEvent | TouchEvent): event is TouchEvent {
-  return Boolean((event as TouchEvent).changedTouches)
-}
